py_utils/requirements.txt
flask==3.1.0
requests==2.32.3
strawberry-graphql[cli]==0.266.0

